Created django project named : backend created app within project named : jc_backend
created model for user, we need a way to map friends to each other, not sure how to implement that
created manager.py, still empty, intended to be a grab-bag of functions
some stuff was automatically generated by PyCharm, which has built in unit testing stuff, dunno if we wanna use that. Just played around with it, to see if I could get it going. Not sure how it works yet